% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klaus-package.R
\docType{package}
\name{klaus-package}
\alias{klaus}
\alias{klaus-package}
\title{klaus: A light-weight package for LLM-based content analysis}
\description{
Provides tools to facilitate automated text classification and content analysis using different LLM provider APIs. Includes functions for API interaction, codebook parsing from CSV, and batch processing of text data according to specified coding schemes.
}
\author{
\strong{Maintainer}: Puschmann Cornelius \email{puschmann@uni-bremen.de}

}
\keyword{internal}
