% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/your_script_name.R
\name{parse_codebook}
\alias{parse_codebook}
\title{Parse a Codebook into JSON}
\usage{
parse_codebook(filepath)
}
\arguments{
\item{filepath}{data.frame. A data frame containing the columns
'category', 'label', and 'instructions'.}
}
\value{
character. An invisible JSON string representing the codebook,
structured with category names as keys.
}
\description{
Reads a data frame defining coding categories, labels, and instructions,
validates its structure, and converts it into a JSON string suitable for
use with the API call. Handles single or multiple categories.
}
\examples{
\dontrun{
codebook <- data.frame(category = c("sentiment", "sentiment", "sentiment"), label = c("positive", "negative", "neutral"), instructions = c("Code this if the sentiment of the tweet is positive", "Code this if the sentiment of the tweet is negative", "Code this if the sentiment of the tweet is neutral"))

json_codebook <- parse_codebook("codebook_content")

cat(json_codebook) # Print the JSON string
}
}
